pyproject.toml
[project]
name = "backend-schoolcollab"
version = "0.1.0"
description = ""
authors = [
    {name = "el hadji",email = "elhadji.sarr@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.3,<0.35.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "tortoise-orm[asyncpg] (>=0.25.1,<0.26.0)",
    "aerich (>=0.9.0,<0.10.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "sentence-transformers (>=3.3.0,<4.0.0)",
    "transformers (>=4.47.0,<5.0.0)",
    "torch (>=2.5.0,<3.0.0)",
    "numpy (>=1.24.0,<2.0.0)",
    "scikit-learn (>=1.3.0,<2.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
